// client/src/app/interfaces/patient.interface.ts
export interface Patient {
    _id: string;
    name: string;
    lastname: string;
    phone: string;
    birthday: Date;
    gender: string;
}
export interface NewPatient {
    name: string;
    lastname: string;
    phone: string;
    birthday: Date;
    gender: string;
}
export interface Patients {
    _id: string;
    name: string;
    lastname: string;
    phone: string;
    birthday: Date;
    gender: string;
}